toolkit/components/extensions/test/mochitest/test_ext_notifications.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for notifications</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript;version=1.8">

add_task(function* test_notifications() {
  function backgroundScript() {
    browser.test.log("running background script");

    var opts = {title: "Testing Notification", message: "Carry on"};

    // Test an unimplemented listener.
    browser.notifications.onClicked.addListener(function() {});

    browser.notifications.create("5", opts, function(id) {
      browser.test.sendMessage("running", id);
      browser.test.notifyPass("background test passed");
    });
  }

  let extensionData = {
    manifest: {
      permissions: ["notifications"]
    },
    background: "(" + backgroundScript.toString() + ")()"
  };

  let extension = ExtensionTestUtils.loadExtension(extensionData);
  info("load complete");
  yield extension.startup();
  info("startup complete");
  let x = yield extension.awaitMessage("running");
  is(x, "5", "got correct value from extension");
  yield extension.awaitFinish();
  info("test complete");
  yield extension.unload();
  info("extension unloaded successfully");
});

</script>

</body>
</html>
